backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
PyMySQL==1.1.1
Pillow==10.4.0
python-decouple==3.8
django-filter==24.3
django-ratelimit==4.1.0
